requirements.txt
faiss_gpu==1.7.2
Flask==2.2.5
numpy==1.24.3
sentence_transformers==2.2.2
torch==2.0.0
